test: use external test packages for NE acceptance tests
`internal/acceptance` imports the `equinix` package to build the
provider under test, so a Network Edge acceptance test that lives in the
same package as the resource it exercises creates an import cycle
(equinix -> networkedge/x -> [test] internal/acceptance -> equinix).

Put the acceptance tests in external `<pkg>_test` packages, which Go
excludes from the cycle, matching what the Fabric resource packages
already do. Unit tests that need package-internal symbols stay in the
internal test package.

The shared device/ACL/SSH config templates move to
`templates_acc_test.go` in each package that needs them.
