+++ internal/resources/networkedge/device/resource_acc_test.go
@@ -1,40 +1,22 @@
-package device
+package device_test
 
 import (
 	"context"
 	"fmt"
 	"log"
+	"maps"
 	"testing"
 
+	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/equinix/terraform-provider-equinix/internal/comparisons"
 	"github.com/equinix/terraform-provider-equinix/internal/config"
-	"github.com/equinix/terraform-provider-equinix/internal/nprintf"
+	"github.com/equinix/terraform-provider-equinix/internal/sweep"
 
 	"github.com/equinix/ne-go"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
-)
-
-const (
-	tstResourcePrefix = "tfacc"
-
-	networkDeviceProjectId                  = "TF_ACC_NETWORK_DEVICE_PROJECT_ID"
-	networkDeviceAccountNameEnvVar          = "TF_ACC_NETWORK_DEVICE_BILLING_ACCOUNT_NAME"
-	networkDeviceSecondaryAccountNameEnvVar = "TF_ACC_NETWORK_DEVICE_SECONDARY_BILLING_ACCOUNT_NAME"
-	networkDeviceMetroEnvVar                = "TF_ACC_NETWORK_DEVICE_METRO"
-	networkDeviceSecondaryMetroEnvVar       = "TF_ACC_NETWORK_DEVICE_SECONDARY_METRO"
-	networkDeviceCSRSDWANLicenseFileEnvVar  = "TF_ACC_NETWORK_DEVICE_CSRSDWAN_LICENSE_FILE"
-	networkDeviceVSRXLicenseFileEnvVar      = "TF_ACC_NETWORK_DEVICE_VSRX_LICENSE_FILE"
-	networkDeviceVersaController1EnvVar     = "TF_ACC_NETWORK_DEVICE_VERSA_CONTROLLER1"
-	networkDeviceVersaController2EnvVar     = "TF_ACC_NETWORK_DEVICE_VERSA_CONTROLLER2"
-	networkDeviceVersaLocalIDEnvVar         = "TF_ACC_NETWORK_DEVICE_VERSA_LOCALID"
-	networkDeviceVersaRemoteIDEnvVar        = "TF_ACC_NETWORK_DEVICE_VERSA_REMOTEID"
-	networkDeviceVersaSerialNumberEnvVar    = "TF_ACC_NETWORK_DEVICE_VERSA_SERIAL"
-	networkDeviceCGENIXLicenseKeyEnvVar     = "TF_ACC_NETWORK_DEVICE_CGENIX_LICENSE_KEY"
-	networkDeviceCGENIXLicenseSecretEnvVar  = "TF_ACC_NETWORK_DEVICE_CGENIX_LICENSE_SECRET"
-	networkDevicePANWLicenseTokenEnvVar     = "TF_ACC_NETWORK_DEVICE_PANW_LICENSE_TOKEN"
 )
 
 func init() {
@@ -45,8 +27,97 @@ func init() {
 	})
 }
 
+func testAccNeSSHUserExists(resourceName string, user *ne.SSHUser) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[resourceName]
+		if !ok {
+			return fmt.Errorf("resource not found: %s", resourceName)
+		}
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("resource has no ID attribute set")
+		}
+		client := acceptance.TestAccProvider.Meta().(*config.Config).Ne
+		resp, err := client.GetSSHUser(rs.Primary.ID)
+		if err != nil {
+			return fmt.Errorf("error when fetching SSH user '%s': %s", rs.Primary.ID, err)
+		}
+		*user = *resp
+		return nil
+	}
+}
+
+func testAccNeSSHUserAttributes(user *ne.SSHUser, devices []*ne.Device, ctx map[string]interface{}) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		if v, ok := ctx["username"]; ok && ne.StringValue(user.Username) != v.(string) {
+			return fmt.Errorf("name does not match %v - %v", ne.StringValue(user.Username), v)
+		}
+		deviceIDs := make([]string, len(devices))
+		for i := range devices {
+			deviceIDs[i] = ne.StringValue(devices[i].UUID)
+		}
+		if !comparisons.SlicesMatch(deviceIDs, user.DeviceUUIDs) {
+			return fmt.Errorf("device_ids does not match %v - %v", deviceIDs, user.DeviceUUIDs)
+		}
+		return nil
+	}
+}
+
+func testAccNetworkACLTemplateExists(resourceName string, template *ne.ACLTemplate) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[resourceName]
+		if !ok {
+			return fmt.Errorf("resource not found: %s", resourceName)
+		}
+		client := acceptance.TestAccProvider.Meta().(*config.Config).Ne
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("resource has no ID attribute set")
+		}
+		resp, err := client.GetACLTemplate(rs.Primary.ID)
+		if err != nil {
+			return fmt.Errorf("error when fetching ACL template '%s': %s", rs.Primary.ID, err)
+		}
+		*template = *resp
+		return nil
+	}
+}
+
+func testAccNetworkACLTemplateAttributes(template *ne.ACLTemplate, ctx map[string]interface{}) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		if v, ok := ctx["name"]; ok && ne.StringValue(template.Name) != v.(string) {
+			return fmt.Errorf("name does not match %v - %v", ne.StringValue(template.Name), v)
+		}
+		if v, ok := ctx["description"]; ok && ne.StringValue(template.Description) != v.(string) {
+			return fmt.Errorf("description does not match %v - %v", ne.StringValue(template.Description), v)
+		}
+		if len(template.InboundRules) != 3 {
+			return fmt.Errorf("number of inbound rules does not match %v - %v", len(template.InboundRules), 3)
+		}
+		for i := 0; i < 3; i++ {
+			if ne.IntValue(template.InboundRules[i].SeqNo) != i+1 {
+				return fmt.Errorf("inbound_rule %d seqNo does not match %v - %v", i+1, ne.IntValue(template.InboundRules[i].SeqNo), i+1)
+			}
+			if v, ok := ctx[fmt.Sprintf("inbound_rule_%d_subnet", i+1)]; ok && ne.StringValue(template.InboundRules[i].Subnet) != v.(string) {
+				return fmt.Errorf("inbound_rule %d subnet does not match %v - %v", i+1, ne.StringValue(template.InboundRules[i].Subnet), v)
+			}
+			if v, ok := ctx[fmt.Sprintf("inbound_rule_%d_protocol", i+1)]; ok && ne.StringValue(template.InboundRules[i].Protocol) != v.(string) {
+				return fmt.Errorf("inbound_rule %d protocol does not match %v - %v", i+1, ne.StringValue(template.InboundRules[i].Protocol), v)
+			}
+			if v, ok := ctx[fmt.Sprintf("inbound_rule_%d_src_port", i+1)]; ok && ne.StringValue(template.InboundRules[i].SrcPort) != v.(string) {
+				return fmt.Errorf("inbound_rule %d src_port does not match %v - %v", i+1, ne.StringValue(template.InboundRules[i].SrcPort), v)
+			}
+			if v, ok := ctx[fmt.Sprintf("inbound_rule_%d_dst_port", i+1)]; ok && ne.StringValue(template.InboundRules[i].DstPort) != v.(string) {
+				return fmt.Errorf("inbound_rule %d dst_port does not match %v - %v", i+1, ne.StringValue(template.InboundRules[i].DstPort), v)
+			}
+			if v, ok := ctx[fmt.Sprintf("inbound_rule_%d_description", i+1)]; ok && ne.StringValue(template.InboundRules[i].Description) != v.(string) {
+				return fmt.Errorf("inbound_rule %d description does not match %v - %v", i+1, ne.StringValue(template.InboundRules[i].Description), v)
+			}
+		}
+		return nil
+	}
+}
+
 func testSweepNetworkDevice(region string) error {
-	config, err := sharedConfigForRegion(region)
+	config, err := sweep.SharedConfigForRegion(region)
 	if err != nil {
 		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping Network devices: %s", err)
 	}
@@ -69,7 +140,7 @@ func testSweepNetworkDevice(region string) error {
 	}
 	nonSweepableCount := 0
 	for _, device := range devices {
-		if !isSweepableTestResource(ne.StringValue(device.Name)) {
+		if !sweep.IsSweepableTestResource(ne.StringValue(device.Name)) {
 			nonSweepableCount++
 			continue
 		}
@@ -92,7 +163,7 @@ func TestAccNetworkDevice_CSR1000V_HA_Managed_Sub(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
 	projectId, _ := schema.EnvDefaultFunc(networkDeviceProjectId, "")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":            "test",
 		"device-account_name":            accountName.(string),
 		"device-self_managed":            false,
@@ -121,7 +192,7 @@ func TestAccNetworkDevice_CSR1000V_HA_Managed_Sub(t *testing.T) {
 		"user-password":                  acctest.RandString(10),
 	}
 
-	contextWithACLs := copyMap(context)
+	contextWithACLs := maps.Clone(context)
 	contextWithACLs["acl-resourceName"] = "acl-pri"
 	contextWithACLs["acl-name"] = fmt.Sprintf("%s-%s", tstResourcePrefix, acctest.RandString(6))
 	contextWithACLs["acl-description"] = acctest.RandString(50)
@@ -136,8 +207,8 @@ func TestAccNetworkDevice_CSR1000V_HA_Managed_Sub(t *testing.T) {
 	var user ne.SSHUser
 	var primaryACL, secondaryACL ne.ACLTemplate
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withSSHUser().build(),
@@ -181,7 +252,7 @@ func TestAccNetworkDevice_CSR1000V_HA_Managed_Sub(t *testing.T) {
 func TestAccNetworkDevice_CSR1000V_HA_Self_BYOL(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":            "test",
 		"device-account_name":            accountName.(string),
 		"device-self_managed":            true,
@@ -207,7 +278,7 @@ func TestAccNetworkDevice_CSR1000V_HA_Self_BYOL(t *testing.T) {
 		"sshkey-public_key":              "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQCXdzXBHaVpKpdO0udnB+4JOgUq7APO2rPXfrevvlZrps98AtlwXXVWZ5duRH5NFNfU4G9HCSiAPsebgjY0fG85tcShpXfHfACLt0tBW8XhfLQP2T6S50FQ1brBdURMDCMsD7duOXqvc0dlbs2/KcswHvuUmqVzob3bz7n1bQ48wIHsPg4ARqYhy5LN3OkllJH/6GEfqi8lKZx01/P/gmJMORcJujuOyXRB+F2iXBVYdhjML3Qg4+tEekBcVZOxUbERRZ0pvQ52Y6wUhn2VsjljixyqeOdmD0m6DayDQgSWms6bKPpBqN7zhXXk4qe8bXT4tQQba65b2CQ2A91jw2KgM/YZNmjyUJ+Rf1cQosJf9twqbAZDZ6rAEmj9zzvQ5vD/CGuzxdVMkePLlUK4VGjPu7cVzhXrnq4318WqZ5/lNiCST8NQ0fssChN8ANUzr/p/wwv3faFMVNmjxXTZMsbMFT/fbb2MVVuqNFN65drntlg6/xEao8gZROuRYiakBx8= user@host",
 	}
 
-	contextWithACLs := copyMap(context)
+	contextWithACLs := maps.Clone(context)
 	contextWithACLs["acl-resourceName"] = "acl-pri"
 	contextWithACLs["acl-name"] = fmt.Sprintf("%s-%s", tstResourcePrefix, acctest.RandString(6))
 	contextWithACLs["acl-description"] = acctest.RandString(50)
@@ -220,8 +291,8 @@ func TestAccNetworkDevice_CSR1000V_HA_Self_BYOL(t *testing.T) {
 	var primary, secondary ne.Device
 	var primaryACL, secondaryACL ne.ACLTemplate
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withSSHKey().build(),
@@ -257,7 +328,7 @@ func TestAccNetworkDevice_CSR1000V_HA_Self_BYOL(t *testing.T) {
 func TestAccNetworkDevice_vSRX_HA_Managed_Sub(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":            "test",
 		"device-account_name":            accountName.(string),
 		"device-self_managed":            false,
@@ -278,7 +349,7 @@ func TestAccNetworkDevice_vSRX_HA_Managed_Sub(t *testing.T) {
 		"device-secondary_notifications": []string{"secondary@equinix.com"},
 	}
 
-	contextWithChanges := copyMap(context)
+	contextWithChanges := maps.Clone(context)
 	contextWithChanges["user-resourceName"] = "test"
 	contextWithChanges["user-username"] = fmt.Sprintf("%s-%s", tstResourcePrefix, acctest.RandString(6))
 	contextWithChanges["user-password"] = acctest.RandString(10)
@@ -287,8 +358,8 @@ func TestAccNetworkDevice_vSRX_HA_Managed_Sub(t *testing.T) {
 	var primary, secondary ne.Device
 	var user ne.SSHUser
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().build(),
@@ -327,7 +398,7 @@ func TestAccNetworkDevice_vSRX_HA_Managed_BYOL(t *testing.T) {
 	if licenseFile.(string) == "" {
 		t.Skip("Skipping TestAccNetworkDevice_vSRX_HA_Managed_BYOL test since TF_ACC_NETWORK_DEVICE_VSRX_LICENSE_FILE env var is not defined with a valid license file")
 	}
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":            "test",
 		"device-account_name":            accountName.(string),
 		"device-self_managed":            false,
@@ -356,7 +427,7 @@ func TestAccNetworkDevice_vSRX_HA_Managed_BYOL(t *testing.T) {
 		"acl-secondary_description":      acctest.RandString(50),
 	}
 
-	contextWithChanges := copyMap(context)
+	contextWithChanges := maps.Clone(context)
 	contextWithChanges["device-name"] = fmt.Sprintf("%s-%s", tstResourcePrefix, acctest.RandString(6))
 	contextWithChanges["device-additional_bandwidth"] = 100
 	contextWithChanges["device-notifications"] = []string{"jerry@equinix.com", "tom@equinix.com"}
@@ -371,8 +442,8 @@ func TestAccNetworkDevice_vSRX_HA_Managed_BYOL(t *testing.T) {
 	var primary, secondary ne.Device
 	var user ne.SSHUser
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withACL().build(),
@@ -409,7 +480,7 @@ func TestAccNetworkDevice_vSRX_HA_Managed_BYOL(t *testing.T) {
 func TestAccNetworkDevice_vSRX_HA_Self_BYOL(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":            "test",
 		"device-account_name":            accountName.(string),
 		"device-self_managed":            true,
@@ -442,8 +513,8 @@ func TestAccNetworkDevice_vSRX_HA_Self_BYOL(t *testing.T) {
 	deviceResourceName := fmt.Sprintf("equinix_network_device.%s", context["device-resourceName"].(string))
 	var primary, secondary ne.Device
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withACL().withSSHKey().build(),
@@ -465,7 +536,7 @@ func TestAccNetworkDevice_vSRX_HA_Self_BYOL(t *testing.T) {
 func TestAccNetworkDevice_PaloAlto_HA_Managed_Sub(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":            "test",
 		"device-account_name":            accountName.(string),
 		"device-self_managed":            false,
@@ -492,7 +563,7 @@ func TestAccNetworkDevice_PaloAlto_HA_Managed_Sub(t *testing.T) {
 		"acl-secondary_description":      acctest.RandString(50),
 	}
 
-	contextWithChanges := copyMap(context)
+	contextWithChanges := maps.Clone(context)
 	contextWithChanges["device-additional_bandwidth"] = 50
 	contextWithChanges["device-secondary_additional_bandwidth"] = 50
 	contextWithChanges["user-resourceName"] = "tst-user"
@@ -506,8 +577,8 @@ func TestAccNetworkDevice_PaloAlto_HA_Managed_Sub(t *testing.T) {
 	secACLResourceName := fmt.Sprintf("equinix_network_acl_template.%s", context["acl-secondary_resourceName"].(string))
 	userResourceName := fmt.Sprintf("equinix_network_ssh_user.%s", contextWithChanges["user-resourceName"].(string))
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withACL().build(),
@@ -548,7 +619,7 @@ func TestAccNetworkDevice_PaloAlto_HA_Managed_Sub(t *testing.T) {
 func TestAccNetworkDevice_PaloAlto_HA_Self_BYOL(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":            "test",
 		"device-account_name":            accountName.(string),
 		"device-self_managed":            true,
@@ -573,7 +644,7 @@ func TestAccNetworkDevice_PaloAlto_HA_Self_BYOL(t *testing.T) {
 		"sshkey-public_key":              "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQCXdzXBHaVpKpdO0udnB+4JOgUq7APO2rPXfrevvlZrps98AtlwXXVWZ5duRH5NFNfU4G9HCSiAPsebgjY0fG85tcShpXfHfACLt0tBW8XhfLQP2T6S50FQ1brBdURMDCMsD7duOXqvc0dlbs2/KcswHvuUmqVzob3bz7n1bQ48wIHsPg4ARqYhy5LN3OkllJH/6GEfqi8lKZx01/P/gmJMORcJujuOyXRB+F2iXBVYdhjML3Qg4+tEekBcVZOxUbERRZ0pvQ52Y6wUhn2VsjljixyqeOdmD0m6DayDQgSWms6bKPpBqN7zhXXk4qe8bXT4tQQba65b2CQ2A91jw2KgM/YZNmjyUJ+Rf1cQosJf9twqbAZDZ6rAEmj9zzvQ5vD/CGuzxdVMkePLlUK4VGjPu7cVzhXrnq4318WqZ5/lNiCST8NQ0fssChN8ANUzr/p/wwv3faFMVNmjxXTZMsbMFT/fbb2MVVuqNFN65drntlg6/xEao8gZROuRYiakBx8= user@host",
 	}
 
-	contextWithACLs := copyMap(context)
+	contextWithACLs := maps.Clone(context)
 	contextWithACLs["acl-resourceName"] = "acl-pri"
 	contextWithACLs["acl-name"] = fmt.Sprintf("%s-%s", tstResourcePrefix, acctest.RandString(6))
 	contextWithACLs["acl-description"] = acctest.RandString(50)
@@ -586,8 +657,8 @@ func TestAccNetworkDevice_PaloAlto_HA_Self_BYOL(t *testing.T) {
 	var primary, secondary ne.Device
 	var primaryACL, secondaryACL ne.ACLTemplate
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withSSHKey().build(),
@@ -623,7 +694,7 @@ func TestAccNetworkDevice_CSRSDWAN_HA_Self_BYOL(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
 	licFile, _ := schema.EnvDefaultFunc(networkDeviceCSRSDWANLicenseFileEnvVar, "test-fixtures/CSRSDWAN.cfg")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":                           "test",
 		"device-account_name":                           accountName.(string),
 		"device-self_managed":                           true,
@@ -664,8 +735,8 @@ func TestAccNetworkDevice_CSRSDWAN_HA_Self_BYOL(t *testing.T) {
 	var primary, secondary ne.Device
 	var primaryACL, secondaryACL ne.ACLTemplate
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withACL().build(),
@@ -700,7 +771,7 @@ func TestAccNetworkDevice_Versa_HA_Self_BYOL(t *testing.T) {
 	localID, _ := schema.EnvDefaultFunc(networkDeviceVersaLocalIDEnvVar, "test@versa.com")()
 	remoteID, _ := schema.EnvDefaultFunc(networkDeviceVersaRemoteIDEnvVar, "test@versa.com")()
 	serialNumber, _ := schema.EnvDefaultFunc(networkDeviceVersaSerialNumberEnvVar, "Test")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":                        "test",
 		"device-account_name":                        accountName.(string),
 		"device-self_managed":                        true,
@@ -743,8 +814,8 @@ func TestAccNetworkDevice_Versa_HA_Self_BYOL(t *testing.T) {
 	var primary, secondary ne.Device
 	var primaryACL, secondaryACL ne.ACLTemplate
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withACL().build(),
@@ -774,7 +845,7 @@ func TestAccNetworkDevice_CGENIX_HA_Self_BYOL(t *testing.T) {
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
 	licenseKey, _ := schema.EnvDefaultFunc(networkDeviceCGENIXLicenseKeyEnvVar, acctest.RandString(10))()
 	licenseSecret, _ := schema.EnvDefaultFunc(networkDeviceCGENIXLicenseSecretEnvVar, acctest.RandString(10))()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":                         "test",
 		"device-account_name":                         accountName.(string),
 		"device-self_managed":                         true,
@@ -811,8 +882,8 @@ func TestAccNetworkDevice_CGENIX_HA_Self_BYOL(t *testing.T) {
 	var primary, secondary ne.Device
 	var primaryACL, secondaryACL ne.ACLTemplate
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withACL().build(),
@@ -841,7 +912,7 @@ func TestAccNetworkDevice_PaloAlto_Cluster_Self_BYOL(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
 	accountName, _ := schema.EnvDefaultFunc(networkDeviceAccountNameEnvVar, "")()
 	licenseToken, _ := schema.EnvDefaultFunc(networkDevicePANWLicenseTokenEnvVar, "")()
-	context := map[string]any{
+	context := map[string]interface{}{
 		"device-resourceName":                "test",
 		"device-account_name":                accountName.(string),
 		"device-self_managed":                true,
@@ -879,8 +950,8 @@ func TestAccNetworkDevice_PaloAlto_Cluster_Self_BYOL(t *testing.T) {
 	var primary ne.Device
 	var wanAcl, mgmtAcl ne.ACLTemplate
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: newTestAccConfig(context).withDevice().withSSHKey().withACL().build(),
@@ -916,7 +987,7 @@ func testAccNeDeviceExists(resourceName string, device *ne.Device) resource.Test
 		if rs.Primary.ID == "" {
 			return fmt.Errorf("resource has no ID attribute set")
 		}
-		client := testAccProvider.Meta().(*config.Config).Ne
+		client := acceptance.TestAccProvider.Meta().(*config.Config).Ne
 		resp, err := client.GetDevice(rs.Primary.ID)
 		if err != nil {
 			return fmt.Errorf("error when fetching network device '%s': %s", rs.Primary.ID, err)
@@ -931,7 +1002,7 @@ func testAccNeDeviceSecondaryExists(primary, secondary *ne.Device) resource.Test
 		if ne.StringValue(primary.RedundantUUID) == "" {
 			return fmt.Errorf("secondary device UUID is not set")
 		}
-		client := testAccProvider.Meta().(*config.Config).Ne
+		client := acceptance.TestAccProvider.Meta().(*config.Config).Ne
 		resp, err := client.GetDevice(ne.StringValue(primary.RedundantUUID))
 		if err != nil {
 			return fmt.Errorf("error when fetching network device '%s': %s", ne.StringValue(primary.RedundantUUID), err)
@@ -950,7 +1021,7 @@ func testAccNeDevicePairExists(resourceName string, primary, secondary *ne.Devic
 		if rs.Primary.ID == "" {
 			return fmt.Errorf("resource has no ID attribute set")
 		}
-		client := testAccProvider.Meta().(*config.Config).Ne
+		client := acceptance.TestAccProvider.Meta().(*config.Config).Ne
 		resp, err := client.GetDevice(rs.Primary.ID)
 		if err != nil {
 			return fmt.Errorf("error when fetching primary network device '%s': %s", rs.Primary.ID, err)
@@ -965,9 +1036,9 @@ func testAccNeDevicePairExists(resourceName string, primary, secondary *ne.Devic
 	}
 }
 
-func testAccNeDeviceAttributes(device *ne.Device, ctx map[string]any) resource.TestCheckFunc {
+func testAccNeDeviceAttributes(device *ne.Device, ctx map[string]interface{}) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
-		if v, ok := ctx["device-name"]; ok && (ne.StringValue(device.Name) != v.(string) && ne.StringValue(device.Name) != v.(string)+"-Node0") {
+		if v, ok := ctx["device-name"]; ok && !(ne.StringValue(device.Name) == v.(string) || ne.StringValue(device.Name) == v.(string)+"-Node0") {
 			return fmt.Errorf("name does not match %v - %v", ne.StringValue(device.Name), v)
 		}
 		if v, ok := ctx["device-self_managed"]; ok && ne.BoolValue(device.IsSelfManaged) != v.(bool) {
@@ -1052,8 +1123,8 @@ func testAccNeDeviceAttributes(device *ne.Device, ctx map[string]any) resource.T
 	}
 }
 
-func testAccNeDeviceSecondaryAttributes(device *ne.Device, ctx map[string]any) resource.TestCheckFunc {
-	secCtx := make(map[string]any)
+func testAccNeDeviceSecondaryAttributes(device *ne.Device, ctx map[string]interface{}) resource.TestCheckFunc {
+	secCtx := make(map[string]interface{})
 	for key, value := range ctx {
 		secCtx[key] = value
 	}
@@ -1140,7 +1211,7 @@ func testAccNeDeviceACL(resourceName string, device *ne.Device) resource.TestChe
 		}
 		templateId := rs.Primary.ID
 		deviceID := ne.StringValue(device.UUID)
-		client := testAccProvider.Meta().(*config.Config).Ne
+		client := acceptance.TestAccProvider.Meta().(*config.Config).Ne
 		if ne.StringValue(device.ACLTemplateUUID) != rs.Primary.ID {
 			return fmt.Errorf("acl_template_id for device %s does not match %v - %v", deviceID, ne.StringValue(device.ACLTemplateUUID), templateId)
 		}
@@ -1182,7 +1253,7 @@ func testAccNeDeviceClusterAttributes(deviceResourceName string) resource.TestCh
 	)
 }
 
-func testAccNeDeviceClusterNodeAttributes(device *ne.Device, ctx map[string]any) resource.TestCheckFunc {
+func testAccNeDeviceClusterNodeAttributes(device *ne.Device, ctx map[string]interface{}) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		cluster := device.ClusterDetails
 		if v, ok := ctx["device-cluster_name"]; ok && ne.StringValue(cluster.ClusterName) != v.(string) {
@@ -1226,393 +1297,4 @@ func testAccNeDeviceClusterNodeAttributes(device *ne.Device, ctx map[string]any)
 		}
 		return nil
 	}
-}
-
-func (t *testAccConfig) withDevice() *testAccConfig {
-	t.config += testAccNetworkDevice(t.ctx)
-	return t
-}
-
-func (t *testAccConfig) withACL() *testAccConfig {
-	t.config += testAccNetworkDeviceACL(t.ctx)
-	return t
-}
-
-func (t *testAccConfig) withSSHKey() *testAccConfig {
-	t.config += testAccNetworkDeviceSSHKey(t.ctx)
-	return t
-}
-
-func testAccNetworkDevice(ctx map[string]any) string {
-	var config string
-	config += nprintf.NPrintf(`
-data "equinix_network_account" "test" {
-  metro_code = "%{device-metro_code}"
-  status     = "Active"
-  project_id = "%{device-project_id}"`, ctx)
-	if v, ok := ctx["device-account_name"]; ok && !comparisons.IsEmpty(v) {
-		config += nprintf.NPrintf(`
-  name = "%{device-account_name}"`, ctx)
-	}
-	config += nprintf.NPrintf(`
-}`, ctx)
-	if _, ok := ctx["device-secondary_metro_code"]; ok {
-		config += nprintf.NPrintf(`
-data "equinix_network_account" "test-secondary" {
-  metro_code = "%{device-secondary_metro_code}"
-  status     = "Active"`, ctx)
-		if v, ok := ctx["device-secondary_account_name"]; ok && !comparisons.IsEmpty(v) {
-			config += nprintf.NPrintf(`
-  name = "%{device-secondary_account_name}"`, ctx)
-		}
-		config += nprintf.NPrintf(` 
-}`, ctx)
-	}
-	config += nprintf.NPrintf(`
-resource "equinix_network_device" "%{device-resourceName}" {
-  self_managed          = %{device-self_managed}
-  byol                  = %{device-byol}
-  name                  = "%{device-name}"
-  metro_code            = "%{device-metro_code}"
-  type_code             = "%{device-type_code}"
-  project_id            = "%{device-project_id}"
-  package_code          = "%{device-package_code}"
-  notifications         = %{device-notifications}
-  term_length           = %{device-term_length}
-  account_number        = data.equinix_network_account.test.number
-  version               = "%{device-version}"
-  core_count            = %{device-core_count}`, ctx)
-	if _, ok := ctx["device-purchase_order_number"]; ok {
-		config += nprintf.NPrintf(`
-  purchase_order_number = "%{device-purchase_order_number}"`, ctx)
-	}
-	if _, ok := ctx["device-purchase_order_number"]; ok {
-		config += nprintf.NPrintf(`
-  order_reference       = "%{device-order_reference}"`, ctx)
-	}
-	if _, ok := ctx["device-additional_bandwidth"]; ok {
-		config += nprintf.NPrintf(`
-  additional_bandwidth  = %{device-additional_bandwidth}`, ctx)
-	}
-	if _, ok := ctx["device-throughput"]; ok {
-		config += nprintf.NPrintf(`
-  throughput            = %{device-throughput}
-  throughput_unit       = "%{device-throughput_unit}"`, ctx)
-	}
-	if _, ok := ctx["device-hostname"]; ok {
-		config += nprintf.NPrintf(`
-  hostname              = "%{device-hostname}"`, ctx)
-	}
-	if _, ok := ctx["device-interface_count"]; ok {
-		config += nprintf.NPrintf(`
-  interface_count       = %{device-interface_count}`, ctx)
-	}
-	if _, ok := ctx["acl-resourceName"]; ok {
-		config += nprintf.NPrintf(`
-  acl_template_id       = equinix_network_acl_template.%{acl-resourceName}.id`, ctx)
-	}
-	if _, ok := ctx["mgmtAcl-resourceName"]; ok {
-		config += nprintf.NPrintf(`
-  mgmt_acl_template_uuid = equinix_network_acl_template.%{mgmtAcl-resourceName}.id`, ctx)
-	}
-	if _, ok := ctx["sshkey-resourceName"]; ok {
-		config += nprintf.NPrintf(`
-  ssh_key {
-    username = "test"
-    key_name = equinix_network_ssh_key.%{sshkey-resourceName}.name
-  }`, ctx)
-	}
-	if _, ok := ctx["device-license_file"]; ok {
-		config += nprintf.NPrintf(`
-  license_file          = "%{device-license_file}"`, ctx)
-	}
-	if _, ok := ctx["device-vendorConfig_enabled"]; ok {
-		config += nprintf.NPrintf(`
-  vendor_configuration  = {`, ctx)
-		if _, ok := ctx["device-vendorConfig_siteId"]; ok {
-			config += nprintf.NPrintf(`
-    siteId          = "%{device-vendorConfig_siteId}"`, ctx)
-		}
-		if _, ok := ctx["device-vendorConfig_systemIpAddress"]; ok {
-			config += nprintf.NPrintf(`
-    systemIpAddress = "%{device-vendorConfig_systemIpAddress}"`, ctx)
-		}
-		if _, ok := ctx["device-vendorConfig_licenseKey"]; ok {
-			config += nprintf.NPrintf(`
-    licenseKey = "%{device-vendorConfig_licenseKey}"`, ctx)
-		}
-		if _, ok := ctx["device-vendorConfig_licenseSecret"]; ok {
-			config += nprintf.NPrintf(`
-    licenseSecret = "%{device-vendorConfig_licenseSecret}"`, ctx)
-		}
-		if _, ok := ctx["device-vendorConfig_controller1"]; ok {
-			config += nprintf.NPrintf(`
-    controller1 = "%{device-vendorConfig_controller1}"`, ctx)
-		}
-		if _, ok := ctx["device-vendorConfig_controller2"]; ok {
-			config += nprintf.NPrintf(`
-    controller2 = "%{device-vendorConfig_controller2}"`, ctx)
-		}
-		if _, ok := ctx["device-vendorConfig_localId"]; ok {
-			config += nprintf.NPrintf(`
-    localId = "%{device-vendorConfig_localId}"`, ctx)
-		}
-		if _, ok := ctx["device-vendorConfig_remoteId"]; ok {
-			config += nprintf.NPrintf(`
-    remoteId = "%{device-vendorConfig_remoteId}"`, ctx)
-		}
-		if _, ok := ctx["device-vendorConfig_serialNumber"]; ok {
-			config += nprintf.NPrintf(`
-    serialNumber = "%{device-vendorConfig_serialNumber}"`, ctx)
-		}
-		config += nprintf.NPrintf(`
-  }`, ctx)
-	}
-	if _, ok := ctx["device-secondary_name"]; ok {
-		config += nprintf.NPrintf(`
-  secondary_device {
-    name                 = "%{device-secondary_name}"`, ctx)
-		if _, ok := ctx["device-secondary_metro_code"]; ok {
-			config += nprintf.NPrintf(`
-    metro_code           = "%{device-secondary_metro_code}"
-    account_number       = data.equinix_network_account.test-secondary.number`, ctx)
-		} else {
-			config += nprintf.NPrintf(`
-    metro_code           = "%{device-metro_code}"
-    account_number       = data.equinix_network_account.test.number`, ctx)
-		}
-		config += nprintf.NPrintf(`
-    notifications        = %{device-secondary_notifications}`, ctx)
-		if _, ok := ctx["device-secondary_additional_bandwidth"]; ok {
-			config += nprintf.NPrintf(`
-    additional_bandwidth = %{device-secondary_additional_bandwidth}`, ctx)
-		}
-		if _, ok := ctx["device-secondary_hostname"]; ok {
-			config += nprintf.NPrintf(`
-    hostname             = "%{device-secondary_hostname}"`, ctx)
-		}
-		if _, ok := ctx["acl-secondary_resourceName"]; ok {
-			config += nprintf.NPrintf(`
-    acl_template_id      = equinix_network_acl_template.%{acl-secondary_resourceName}.id`, ctx)
-		}
-		if _, ok := ctx["mgmtAcl-secondary_resourceName"]; ok {
-			config += nprintf.NPrintf(`
-    mgmt_acl_template_uuid = equinix_network_acl_template.%{mgmtAcl-secondary_resourceName}.id`, ctx)
-		}
-		if _, ok := ctx["sshkey-resourceName"]; ok {
-			config += nprintf.NPrintf(`
-    ssh_key {
-      username = "test"
-      key_name = equinix_network_ssh_key.%{sshkey-resourceName}.name
-    }`, ctx)
-		}
-		if _, ok := ctx["device-secondary_license_file"]; ok {
-			config += nprintf.NPrintf(`
-    license_file         = "%{device-secondary_license_file}"`, ctx)
-		}
-		if _, ok := ctx["device-secondary_vendorConfig_enabled"]; ok {
-			config += nprintf.NPrintf(`
-    vendor_configuration  = {`, ctx)
-			if _, ok := ctx["device-secondary_vendorConfig_siteId"]; ok {
-				config += nprintf.NPrintf(`
-      siteId          = "%{device-secondary_vendorConfig_siteId}"`, ctx)
-			}
-			if _, ok := ctx["device-secondary_vendorConfig_systemIpAddress"]; ok {
-				config += nprintf.NPrintf(`
-      systemIpAddress = "%{device-secondary_vendorConfig_systemIpAddress}"`, ctx)
-			}
-			if _, ok := ctx["device-secondary_vendorConfig_licenseKey"]; ok {
-				config += nprintf.NPrintf(`
-      licenseKey = "%{device-secondary_vendorConfig_licenseKey}"`, ctx)
-			}
-			if _, ok := ctx["device-secondary_vendorConfig_licenseSecret"]; ok {
-				config += nprintf.NPrintf(`
-      licenseSecret = "%{device-secondary_vendorConfig_licenseSecret}"`, ctx)
-			}
-			if _, ok := ctx["device-secondary_vendorConfig_controller1"]; ok {
-				config += nprintf.NPrintf(`
-      controller1 = "%{device-secondary_vendorConfig_controller1}"`, ctx)
-			}
-			if _, ok := ctx["device-secondary_vendorConfig_controller2"]; ok {
-				config += nprintf.NPrintf(`
-      controller2 = "%{device-secondary_vendorConfig_controller2}"`, ctx)
-			}
-			if _, ok := ctx["device-secondary_vendorConfig_localId"]; ok {
-				config += nprintf.NPrintf(`
-      localId = "%{device-secondary_vendorConfig_localId}"`, ctx)
-			}
-			if _, ok := ctx["device-secondary_vendorConfig_remoteId"]; ok {
-				config += nprintf.NPrintf(`
-      remoteId = "%{device-secondary_vendorConfig_remoteId}"`, ctx)
-			}
-			if _, ok := ctx["device-secondary_vendorConfig_serialNumber"]; ok {
-				config += nprintf.NPrintf(`
-      serialNumber = "%{device-secondary_vendorConfig_serialNumber}"`, ctx)
-			}
-			config += nprintf.NPrintf(`
-    }`, ctx)
-		}
-		config += `
-  }`
-	}
-	if _, ok := ctx["device-cluster_name"]; ok {
-		config += nprintf.NPrintf(`
-  cluster_details {
-    cluster_name        = "%{device-cluster_name}"`, ctx)
-		config += `
-    node0 {`
-		if _, ok := ctx["device-node0_license_file_id"]; ok {
-			config += nprintf.NPrintf(`
-      license_file_id   = "%{device-node0_license_file_id}"`, ctx)
-		}
-		if _, ok := ctx["device-node0_license_token"]; ok {
-			config += nprintf.NPrintf(`
-      license_token     = "%{device-node0_license_token}"`, ctx)
-		}
-		if _, ok := ctx["device-node0_vendorConfig_enabled"]; ok {
-			config += nprintf.NPrintf(`
-      vendor_configuration {`, ctx)
-			if _, ok := ctx["device-node0_vendorConfig_hostname"]; ok {
-				config += nprintf.NPrintf(`
-        hostname        = "%{device-node0_vendorConfig_hostname}"`, ctx)
-			}
-			if _, ok := ctx["device-node0_vendorConfig_adminPassword"]; ok {
-				config += nprintf.NPrintf(`
-        admin_password  = "%{device-node0_vendorConfig_adminPassword}"`, ctx)
-			}
-			if _, ok := ctx["device-node0_vendorConfig_controller1"]; ok {
-				config += nprintf.NPrintf(`
-        controller1     = "%{device-node0_vendorConfig_controller1}"`, ctx)
-			}
-			if _, ok := ctx["device-node0_vendorConfig_activationKey"]; ok {
-				config += nprintf.NPrintf(`
-        activation_key  = "%{device-node0_vendorConfig_activationKey}"`, ctx)
-			}
-			if _, ok := ctx["device-node0_vendorConfig_controllerFqdn"]; ok {
-				config += nprintf.NPrintf(`
-        controller_fqdn = "%{device-node0_vendorConfig_controllerFqdn}"`, ctx)
-			}
-			if _, ok := ctx["device-node0_vendorConfig_rootPassword"]; ok {
-				config += nprintf.NPrintf(`
-        root_password   = "%{device-node0_vendorConfig_rootPassword}"`, ctx)
-			}
-			config += nprintf.NPrintf(`
-      }`, ctx)
-		}
-		config += `
-    }`
-		config += `
-    node1 {`
-		if _, ok := ctx["device-node1_license_file_id"]; ok {
-			config += nprintf.NPrintf(`
-      license_file_id   = "%{device-node1_license_file_id}"`, ctx)
-		}
-		if _, ok := ctx["device-node1_license_token"]; ok {
-			config += nprintf.NPrintf(`
-      license_token     = "%{device-node1_license_token}"`, ctx)
-		}
-		if _, ok := ctx["device-node1_vendorConfig_enabled"]; ok {
-			config += nprintf.NPrintf(`
-      vendor_configuration {`, ctx)
-			if _, ok := ctx["device-node1_vendorConfig_hostname"]; ok {
-				config += nprintf.NPrintf(`
-        hostname        = "%{device-node1_vendorConfig_hostname}"`, ctx)
-			}
-			if _, ok := ctx["device-node1_vendorConfig_adminPassword"]; ok {
-				config += nprintf.NPrintf(`
-        admin_password  = "%{device-node1_vendorConfig_adminPassword}"`, ctx)
-			}
-			if _, ok := ctx["device-node1_vendorConfig_controller1"]; ok {
-				config += nprintf.NPrintf(`
-        controller1     = "%{device-node1_vendorConfig_controller1}"`, ctx)
-			}
-			if _, ok := ctx["device-node1_vendorConfig_activationKey"]; ok {
-				config += nprintf.NPrintf(`
-        activation_key  = "%{device-node1_vendorConfig_activationKey}"`, ctx)
-			}
-			if _, ok := ctx["device-node1_vendorConfig_controllerFqdn"]; ok {
-				config += nprintf.NPrintf(`
-        controller_fqdn = "%{device-node1_vendorConfig_controllerFqdn}"`, ctx)
-			}
-			if _, ok := ctx["device-node1_vendorConfig_rootPassword"]; ok {
-				config += nprintf.NPrintf(`
-        root_password   = "%{device-node1_vendorConfig_rootPassword}"`, ctx)
-			}
-			config += nprintf.NPrintf(`
-      }`, ctx)
-		}
-		config += `
-    }`
-		config += `
-  }`
-	}
-	config += `
-}`
-	return config
-}
-
-func testAccNetworkDeviceACL(ctx map[string]any) string {
-	var config string
-	if _, ok := ctx["acl-name"]; ok {
-		config += nprintf.NPrintf(`
-resource "equinix_network_acl_template" "%{acl-resourceName}" {
-  name          = "%{acl-name}"
-  description   = "%{acl-description}"
-  inbound_rule {
-    subnet   = "10.0.0.0/24"
-    protocol = "IP"
-    src_port = "any"
-    dst_port = "any"
-  }
-}`, ctx)
-	}
-	if _, ok := ctx["mgmtAcl-name"]; ok {
-		config += nprintf.NPrintf(`
-resource "equinix_network_acl_template" "%{mgmtAcl-resourceName}" {
-  name          = "%{mgmtAcl-name}"
-  description   = "%{mgmtAcl-description}"
-  inbound_rule {
-    subnet   = "11.0.0.0/24"
-    protocol = "IP"
-    src_port = "any"
-    dst_port = "any"
-  }
-}`, ctx)
-	}
-	if _, ok := ctx["acl-secondary_name"]; ok {
-		config += nprintf.NPrintf(`
-resource "equinix_network_acl_template" "%{acl-secondary_resourceName}" {
-  name          = "%{acl-secondary_name}"
-  description   = "%{acl-secondary_description}"
-  inbound_rule {
-    subnet   = "192.0.0.0/24"
-    protocol = "IP"
-    src_port = "any"
-    dst_port = "any"
-  }
-}`, ctx)
-	}
-	if _, ok := ctx["mgmtAcl-secondary_name"]; ok {
-		config += nprintf.NPrintf(`
-resource "equinix_network_acl_template" "%{mgmtAcl-secondary_resourceName}" {
-  name          = "%{mgmtAcl-secondary_name}"
-  description   = "%{mgmtAcl-secondary_description}"
-  inbound_rule {
-    subnet   = "193.0.0.0/24"
-    protocol = "IP"
-    src_port = "any"
-    dst_port = "any"
-  }
-}`, ctx)
-	}
-	return config
-}
-
-func testAccNetworkDeviceSSHKey(ctx map[string]any) string {
-	return nprintf.NPrintf(`
-resource "equinix_network_ssh_key" "%{sshkey-resourceName}" {
-  name       = "%{sshkey-name}"
-  public_key = "%{sshkey-public_key}"
-}
-`, ctx)
 }
