+++ internal/resources/networkedge/device_software/data_source_acc_test.go
@@ -1,9 +1,10 @@
-package device_software
+package device_software_test
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/equinix/terraform-provider-equinix/internal/nprintf"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
@@ -18,8 +19,8 @@ func TestAccDataSourceNetworkDeviceSoftware_versionRegex(t *testing.T) {
 	}
 	resourceName := fmt.Sprintf("data.equinix_network_device_software.%s", context["resourceName"].(string))
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccDataSourceNetworkDeviceSoftwareConfig_versionRegex(context),
