+++ internal/resources/networkedge/device_type/data_source_acc_test.go
@@ -1,13 +1,16 @@
-package device_type
+package device_type_test
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/equinix/terraform-provider-equinix/internal/nprintf"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
+
+const networkDeviceMetroEnvVar = "TF_ACC_NETWORK_DEVICE_METRO"
 
 func TestAccDataSourceNetworkDeviceType_basic(t *testing.T) {
 	metro, _ := schema.EnvDefaultFunc(networkDeviceMetroEnvVar, "SV")()
@@ -19,8 +22,8 @@ func TestAccDataSourceNetworkDeviceType_basic(t *testing.T) {
 	}
 	resourceName := fmt.Sprintf("data.equinix_network_device_type.%s", context["resourceName"].(string))
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccDataSourceNetworkDeviceTypeConfig_basic(context),
