+++ internal/resources/networkedge/file/resource_acc_test.go
@@ -1,9 +1,10 @@
-package file
+package file_test
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/equinix/terraform-provider-equinix/internal/config"
 	"github.com/equinix/terraform-provider-equinix/internal/nprintf"
 
@@ -12,6 +13,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
+
+const tstResourcePrefix = "tfacc"
 
 func TestAccNetworkFile_VSRX(t *testing.T) {
 	context := map[string]any{
@@ -27,8 +30,8 @@ func TestAccNetworkFile_VSRX(t *testing.T) {
 	resourceName := fmt.Sprintf("equinix_network_file.%s", context["resourceName"].(string))
 	var file ne.File
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:  func() { acceptance.TestAccPreCheck(t) },
+		Providers: acceptance.TestAccProviders,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccNetworkFile(context),
@@ -69,7 +72,7 @@ func testAccNetworkFileExists(resourceName string, file *ne.File) resource.TestC
 		if !ok {
 			return fmt.Errorf("resource not found: %s", resourceName)
 		}
-		client := testAccProvider.Meta().(*config.Config).Ne
+		client := acceptance.TestAccProvider.Meta().(*config.Config).Ne
 		if rs.Primary.ID == "" {
 			return fmt.Errorf("resource has no ID attribute set")
 		}
